--- MulticapaRESTClient/.classpath
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<classpath>
-	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER"/>
-	<classpathentry kind="lib" path="/home/administrador/Descargas/apache-tomcat-8.5.59/lib/servlet-api.jar">
-		<attributes>
-			<attribute name="javadoc_location" value="https://docs.oracle.com/javaee/7/api/"/>
-		</attributes>
-	</classpathentry>
-	<classpathentry kind="lib" path="/home/administrador/Descargas/javax.ws.rs-api-2.0.1.jar"/>
-	<classpathentry kind="output" path="bin"/>
-</classpath>
