--- MulticapaRESTClient/bin/.gitignore
@@ -1,2 +1,0 @@
-/main/
-/negocio/
