--- RESTPlano/.settings/org.eclipse.wst.common.component
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?><project-modules id="moduleCoreId" project-version="1.5.0">
-    <wb-module deploy-name="RESTPlano">
-        <wb-resource deploy-path="/" source-path="/WebContent" tag="defaultRootSource"/>
-        <wb-resource deploy-path="/WEB-INF/classes" source-path="/src"/>
-        <property name="context-root" value="RESTPlano"/>
-        <property name="java-output-path" value="/RESTPlano/build/classes"/>
-    </wb-module>
-</project-modules>
