--- RESTPlano/WebContent/WEB-INF/web.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<web-app xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://java.sun.com/xml/ns/javaee" xmlns:web="http://java.sun.com/xml/ns/javaee/web-app_2_5.xsd" xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_2_5.xsd" id="WebApp_ID" version="2.5">
-  <display-name>practica</display-name>
- 
-  <servlet>
-    <servlet-name>ServicioServlet</servlet-name>
-    <servlet-class>servicio.Saluda</servlet-class>
-  </servlet>
-  <servlet-mapping>
-    <servlet-name>ServicioServlet</servlet-name>
-    <url-pattern>/saluda</url-pattern>
-  </servlet-mapping>
-</web-app>
