--- RESTPlano/src/negocio/imp/FactoriaNegocioImp.java
@@ -1,15 +1,0 @@
-package negocio.imp;
-
-import negocio.FactoriaNegocio;
-import negocio.saludo.Saludo;
-import negocio.saludo.imp.SaludoImp;
-
-public class FactoriaNegocioImp extends FactoriaNegocio {
-
-	
-	public Saludo nuevoSaludo() {
-		
-		return new SaludoImp();
-	}
-
-}
