--- RESTPlano/src/negocio/saludo/Saludo.java
@@ -1,8 +1,0 @@
-package negocio.saludo;
-
-public interface Saludo {
-	
-	public String saludar(String nombre, String apellido);
-	
-
-}
