--- RESTPlano/src/negocio/saludo/imp/SaludoImp.java
@@ -1,12 +1,0 @@
-package negocio.saludo.imp;
-
-import negocio.saludo.Saludo;
-
-public class SaludoImp implements Saludo {
-	
-	public String saludar(String nombre, String apellido)
-	{
-		return "Hola "+nombre+" "+apellido;
-	}
-
-}
