+++ saludoREST cliente/.settings/org.eclipse.wst.common.project.facet.core.prefs.xml
@@ -1,0 +1,7 @@
+<root>
+  <facet id="jst.jaxrs">
+    <node name="libprov">
+      <attribute name="provider-id" value="jaxrs-user-library-provider"/>
+    </node>
+  </facet>
+</root>
