+++ saludoREST cliente/.settings/org.eclipse.wst.common.project.facet.core.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<faceted-project>
+  <installed facet="java" version="1.8"/>
+  <installed facet="jst.jaxrs" version="1.1"/>
+</faceted-project>
