+++ saludoREST/.settings/org.eclipse.jst.ws.cxf.core.prefs
@@ -1,0 +1,2 @@
+eclipse.preferences.version=1
+org.eclipse.jst.ws.cxf.core.runtime.version=3.2.0
