+++ saludoREST/WebContent/WEB-INF/web.xml
@@ -1,0 +1,41 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<web-app xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://java.sun.com/xml/ns/javaee" xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_3_0.xsd" version="3.0">
+  <display-name>saludoREST</display-name>
+   
+ <!-- ¡Borrar/ignorar esto que es lo que genera Eclipse!
+ <servlet>
+    <description>JAX-RS Tools Generated - Do not modify</description>
+    <servlet-name>JAX-RS Servlet</servlet-name>
+    <load-on-startup>1</load-on-startup>
+  </servlet>
+  <servlet-mapping>
+    <servlet-name>JAX-RS Servlet</servlet-name>
+    <url-pattern>/jaxrs/*</url-pattern>
+  </servlet-mapping>
+ 
+  
+-->
+
+
+ <servlet>
+ <servlet-name>CXFServlet</servlet-name>
+  <servlet-class>
+   org.apache.cxf.jaxrs.servlet.CXFNonSpringJaxrsServlet
+ </servlet-class>
+ <init-param>
+  <param-name>jaxrs.serviceClasses</param-name>
+  <param-value>
+    negocio.localidad.wsb.SALocalidadWSB
+  </param-value>
+ </init-param>
+ 
+ <load-on-startup>1</load-on-startup>
+</servlet>
+
+<servlet-mapping>
+    <servlet-name>CXFServlet</servlet-name>
+    <url-pattern>/servicios/*</url-pattern>
+  </servlet-mapping>
+  
+  
+</web-app>
