+++ saludoREST/src/negocio/imp/FactoriaNegocioImp.java
@@ -1,0 +1,15 @@
+package negocio.imp;
+
+import negocio.FactoriaNegocio;
+import negocio.localidad.SALocalidad;
+import negocio.localidad.imp.SALocalidadImp;
+
+public class FactoriaNegocioImp extends FactoriaNegocio {
+
+	
+	public SALocalidad generaSALocalidad() {
+
+		return new SALocalidadImp();
+	}
+
+}
